refatora os bagulhos
