+++ .env.example
@@ -1,0 +1,1 @@
+GITHUB_TOKEN=
