+++ package.json
@@ -10,6 +10,7 @@
   "type": "module",
   "license": "ISC",
   "dependencies": {
+    "dotenv": "^16.0.1",
     "graphql-request": "^4.3.0",
     "objects-to-csv": "^1.3.6"
   }
